requirements.txt
streamlit
biopython
scikit-learn
pandas

